Fixed resource amount to float

+++ app/Http/Resources/SpinWinningResource.php
@@ -17,7 +17,7 @@ class SpinWinningResource extends JsonResource
     public function toArray(Request $request): array
     {
         return [
-            'amount'    => in_array($this->type, [Spin::FIVE_X, Spin::TEN_X]) ? 5 : 2.5,
+            'amount'    => (float) in_array($this->type, [Spin::FIVE_X, Spin::TEN_X]) ? 5 : 2.5,
             'type'      => $this->spin_type,
             'created_at'=> Carbon::parse($this->created_at)->toDateTimeString()
         ];
